+++ tests/crossover.rs
@@ -15,16 +15,16 @@ use rand_core::OsRng;
 fn crossover_hash() -> Result<(), Error> {
     let mut rng = OsRng;
 
-    let ssk = SecretKey::random(&mut rng);
-    let psk = PublicKey::from(ssk);
+    let sk = SecretKey::random(&mut rng);
+    let pk = PublicKey::from(sk);
 
     let value = 25;
     let blinding_factor = JubJubScalar::random(&mut rng);
-    let note = Note::obfuscated(&mut rng, &psk, value, blinding_factor);
+    let note = Note::obfuscated(&mut rng, &pk, value, blinding_factor);
 
     let value = 25;
     let blinding_factor = JubJubScalar::random(&mut rng);
-    let note_p = Note::obfuscated(&mut rng, &psk, value, blinding_factor);
+    let note_p = Note::obfuscated(&mut rng, &pk, value, blinding_factor);
 
     let (_, crossover) = note.try_into()?;
     let (_, crossover_p) = note_p.try_into()?;